% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_forecasting_functions.R
\name{combine_subspecies}
\alias{combine_subspecies}
\title{Combine subspecies into their common species}
\usage{
combine_subspecies(df)
}
\description{
Combine subspecies into their common species
}
