% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs_forecasting_functions.R
\name{get_env_data}
\alias{get_env_data}
\title{Master function for acquiring all environmental in a single table.
Returns either past observed data from PRSISM + NDVI, or future
forecasted data from CMIP5 and a naive NDVI forecast.}
\usage{
get_env_data(timeframe = "past", ndvi_forecast_source_years = 2000:2013,
  future_data_years = 2014:2050)
}
\arguments{
\item{timeframe}{str Either 'past' or 'future' for 1981-2013 or 2014-2050, respectively.}

\item{ndvi_forecast_source_years}{vector Years of observed NDVI data to use for making 
a naive NDVI forecast with site specific means. Ignored if timeframe is 'past'}

\item{future_data_years}{vector Years of forecast environmental data to return. Ignored
if timeframe is 'past'}
}
\description{
Master function for acquiring all environmental in a single table.
Returns either past observed data from PRSISM + NDVI, or future
forecasted data from CMIP5 and a naive NDVI forecast.
}
